% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rose_Wrapper.R
\name{json_to_df}
\alias{json_to_df}
\title{A data conversion Function}
\usage{
json_to_df(dataset_json)
}
\arguments{
\item{dataset_json}{dataset in json to be converted}
}
\description{
Convert json dataset to dataframes
}
\keyword{internal}
