% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rose_Wrapper.R
\name{rose_login}
\alias{rose_login}
\title{A Rose Login Function}
\usage{
rose_login(username, password)
}
\arguments{
\item{username}{Email of the user}

\item{password}{Password of the user}
}
\description{
Login Rose given an email and password
}
