% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rose_Wrapper.R
\name{rose_pull}
\alias{rose_pull}
\title{A data extraction Function}
\usage{
rose_pull(code, as_dataframe = TRUE, exact_match = TRUE)
}
\arguments{
\item{code}{Code to pull from ROSE}

\item{as_dataframe}{Boolean to determine return object type}

\item{exact_match}{Boolean to determine code match type in Rose}
}
\description{
Pulls the dataset object given a code
}
\examples{
pull("chn.agriculture.pgdp.wdi")
}
