% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rose_Wrapper.R
\name{rose_pull_logic}
\alias{rose_pull_logic}
\title{A logic extraction Function}
\usage{
rose_pull_logic(code)
}
\arguments{
\item{code}{Code to pull from ROSE}
}
\description{
Pulls the logic given a code
}
\examples{
pull_logic(rosecode)
}
