% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rose_Wrapper.R
\name{rose_push}
\alias{rose_push}
\title{A function to push data to Rose}
\usage{
rose_push(code = NULL, metas = NULL, values = NULL,
  overwrite = FALSE, data_type = "timeseries")
}
\arguments{
\item{code}{Code of the dataset}

\item{metas}{Metas dataframe with index as tag names and a single column for tag values}

\item{values}{Values dataframe with index as dates in 'yyyy-mm-dd' format and a single column for values for each timestamp}

\item{data_type}{type of the data, "timeseries" or "map"}
}
\description{
Creates/updates a dataset in Rose
}
