% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rose_Wrapper.R
\name{rose_push_logic}
\alias{rose_push_logic}
\title{Push logic function}
\usage{
rose_push_logic(code, logic)
}
\arguments{
\item{code}{Code to push to ROSE}

\item{logic}{logic for this code}
}
\description{
Push the logic given a code
}
\examples{
rose_push_logic(rosecode, logic)
}
