% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{SignificantIntervalSearchFastCmh-class}
\alias{SignificantIntervalSearchFastCmh-class}
\alias{SignificantIntervalSearchFastCmh}
\title{Fast significant interval search with categorical covariates}
\format{\code{\link{RefClass}} object.}
\description{
Class for significant intervals search with Cochran-Mantel-Haenszel test,
which accounts for categorical covariates.
}

\section{Base methods}{


\describe{
  \item{\code{new()}}{Create new instance of this class. Can be called implicitly as \code{ThisSubclassName()}.}

  \item{\code{read_eth_files(data_path, labels_path)}}{Read the ETH format files (data and labels).}
  \item{\code{read_plink_files(base_path)}}{Read the PLINK format files (with the same \code{base_path} base name).}

  \item{\code{write_eth_files(data_path, labels_path)}}{Write the ETH format files (data and labels).}

  \item{\code{set_alpha(alpha)}}{Set the significance threshold (FWER).}
  \item{\code{get_alpha(alpha)}}{Get the significance threshold (FWER; \code{NULL} if not set).}
  \item{\code{set_lmax(lmax)}}{Set the maximum number of SNPs to consider (\code{0} means consider all).}
  \item{\code{get_lmax(lmax)}}{Get the maximum number of SNPs to consider (\code{0} means consider all; ; \code{NULL} if not set).}

  \item{\code{execute()}}{Execute the search.}

  \item{\code{get_result()}}{Return all results of the search.}

  \item{\code{write_summary(file_path)}}{Write search summary to a file.}
  \item{\code{write_profile(file_path)}}{Write profiling summary to a file.}
}
}

\section{Interval methods}{


\describe{

  \item{\code{get_significant_intervals()}}{Returns found significant intervals.}
  \item{\code{get_filtered_intervals()}}{Returns found significant intervals clustered by overlaps.}

  \item{\code{write_filtered_intervals(file_path)}}{Write filtered intervals (most significant from each cluster) to a file.}
  \item{\code{write_pvals_testable_intervals(file_path)}}{Write p-values for all testable intervals found to a file.}
  \item{\code{write_pvals_significant_intervals(file_path)}}{Write p-values for all significant intervals found to a file.}
}
}

\section{Covariates methods}{


\describe{
  \item{\code{read_eth_files(..., cov_path=NULL)}}{Read the ETH format files together with the covariates file (by default one covariate for all observations).}
  \item{\code{read_plink_files(..., cov_path=NULL)}}{Read the PLINK format files together with the covariates file (by default one covariate for all observations).}
  \item{\code{update_covariates_file(cov_path)}}{Update covariates file.}

  \item{\code{write_eth_files(..., cov_path=NULL)}}{Write the ETH format files together with the covariates file (by default do not write covariates).}
}
}

