% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{sigpatsearch}
\alias{sigpatsearch}
\title{A constructor for a significant pattern search object}
\usage{
sigpatsearch(method = "", use_intervals = NA, use_combinations = NA,
  use_covariate = FALSE, alpha = 0.05, max_length = 0)
}
\arguments{
\item{method}{A string specifying the method. Use either \code{'fais'}, 
\code{'fastcmh'} or \code{'facs'}. Default is code{''}.
Another way to set the method is to set the 
\code{use_intervals}, \code{use_combinations} and
\code{use_covariate} boolean flags (see below).}

\item{use_intervals}{Boolean flag to set whether to use intervals
or combinations. Default is \code{NA}. 
At least one of 'use_intervals' or 
'use_combinations' needs to be set to \code{TRUE}
or \code{FALSE}.}

\item{use_combinations}{Boolean flag to set whether to use combinations 
or intervals Default is \code{NA}.
At least one of 'use_intervals' or 
'use_combinations' needs to be set to \code{TRUE}
or \code{FALSE}.}

\item{use_covariate}{Boolean flag to set whether or not to use a covariate.
Default is \code{NA}, which is interpreted as 
\code{FALSE}.}

\item{alpha}{Value in range \eqn{(0, 1)} (excluding 0 and 1) which sets
significance threshold. If outside range, error is thrown.
Default value is \code{0.05}.}

\item{max_length}{A value which, for the interval-based methods \code{FAIS} 
\code{FastCMH}, if the user so desired, 
sets the maximum interval length. For 
example, if \code{max_length=5}, only intervals up to
length 5 will be considered. Negative values will be 
interpreted as \code{0},
which means that there is no restriction on the interval 
length (all possible intervals will be considered).
Non-integer values will be rounded down to nearest 
integer. Default value is \code{0}}
}
\description{
Initialises an object for the \code{FAIS}, \code{FastCMH} or \code{FACS}
methods. Returns the object.
}
